+++ src/model/package.ts
@@ -21,7 +21,10 @@ export type QueryOptions = object & BaseQueryOptions;
 
 // A regexp that matches paths to external code.
 // TODO(rictic): Make this extensible (polymer.json?).
-const MATCHES_EXTERNAL = /(^|\/)(bower_components|node_modules)/;
+// Note that we match directories named exactly `build`, but will match any
+// directory name prefixed by `bower_components` or `node_modules`, in order to
+// ignore `polymer install`'s variants, which look like bower_components-foo
+const MATCHES_EXTERNAL = /(^|\/)(bower_components|node_modules|build($|\/))/;
 
 /**
  * Represents a queryable interface over all documents in a package/project.